class split to two smaller classes
